OMAP3+: xSI: Fixed incorrect FIFO ioctl return type
The IOCTL for getting FIFO occupancy returns a u8 data type (unsigned char), which
is not very portable.

So changed the returned type from u8 to size_t.

+++ drivers/staging/omap_hsi/hsi_driver_if.c
@@ -800,7 +800,7 @@ int hsi_ioctl(struct hsi_device *dev, unsigned int command, void *arg)
 			err = -EFAULT;
 			goto out;
 		}
-		*(u8 *)arg = hsi_get_rx_fifo_occupancy(hsi_ctrl, fifo);
+		*(size_t *)arg = hsi_get_rx_fifo_occupancy(hsi_ctrl, fifo);
 		break;
 	default:
 		err = -ENOIOCTLCMD;
